Minor fix, changing int to size_t [skip ci]

+++ RibosoftAlgo/src/functions.h
@@ -21,7 +21,7 @@ extern "C" DLL_PUBLIC int math_add(int a, int b);
 extern "C" DLL_PUBLIC R_STATUS validate_sequence(const char* sequence);
 extern "C" DLL_PUBLIC R_STATUS validate_structure(const char* structure);
 
-extern "C" DLL_PUBLIC R_STATUS fold(const char* sequence, /*out*/ fold_output*& output, /*out*/ int& size);
+extern "C" DLL_PUBLIC R_STATUS fold(const char* sequence, /*out*/ fold_output*& output, /*out*/ size_t& size);
 extern "C" DLL_PUBLIC R_STATUS structure(const char* candidate, const char* ideal, /*out*/ float& distance);
 
 RIBOSOFT_NAMESPACE_END
